Figure Factory
The task
There is a factory of figures: a circle, a square, a triangle.
Each shape has parameters: name, color (selected randomly from an array of colors), perimeter and area
When the user starts the program, he should choose one of 4 menu items: 1-circle, 2-square, 3-triangle, 0-exit.
"1. You have selected *figure_name*. Enter the radius / side / sides (depending on the method of counting the perimeter and the area of the triangle):"
The shape parameters and menu should appear.
The program prompts to select a shape while the user presses the exit.